feat: Consolidate employee creation endpoint and update API documentation
- Merge `/api/v1/employees/add` route into main `/api/v1/employees` route
- Add comprehensive error handling and validation for employee creation
- Update API documentation to reflect new endpoint structure
- Improve response formatting with consistent success/error patterns
- Add detailed validation error responses in documentation
